# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: TriReservas API
  description: |-
    TriReservas es una aplicación web para la gestión de alojamientos y reservas, con roles diferenciados para 
    usuarios y anfitriones. Desarrollada con Spring Boot, JPA y Angular, ofrece una solución moderna y eficiente para la
    administración de estancias.
    
    Requisitos Funcionales:
    * [Aquí](https://docs.google.com/spreadsheets/d/14pBo766cm17JRFVoYanh_tZuF3VmeBVc5iNh8Avp4dE/edit?usp=sharing) podrá 
    encontrar un apoyo en hojas de cálculo para la identificación de requisitos funcionales y el progreso del proyecto.

    Autores: 
    * Juan Sebastián Londoño Ramírez [github](https://github.com/Juan3bitas)
    * Angie Carolina Grajales Giraldo [github](https://github.com/angiec-grajalesg)
    * Juan Diego García Nieto [github](https://github.com/Juan-Devgo)
  version: 1.1.1
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
  /api/users:
    get:
      tags:
        - User
      summary: Obtener todos los usuarios
      description: Obtiene una lista de todos los usuarios registrados en el sistema.
      operationId: getUsers
      responses:
        '200':
          description: Usuarios obtenidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Usuarios obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "getUsers401", message: "No está autorizado para ver la lista de usuarios."}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "getUsers403", message: "No tiene permisos para ver la lista de usuarios."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-01", message: "Error inesperado al tratar de obtener los usuarios."}
      security:
        - bearerAuth: []
  /api/users/{user_uuid}:
    get:
      tags:
        - User
      summary: Obtener un usuario
      description: Obtiene un usuario específico identificado por su UUID.
      operationId: getUser
      responses:
        '200':
          description: Usuario obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Usuario obtenido satisfactoriamente."
                  content:
                    $ref: '#/components/schemas/UserDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: { code: "getUser401", message: "No autorizado para esta acción. (JWT ausente)" }
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: { code: "getUser403", message: "Prohibido para esta acción (JWT inválido o sin permisos)" }
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "getUser404", message: "No se encontró el usuario especificado para obtener."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-26", message: "Error inesperado al tratar de obtener un usuario."}
      security:
        - bearerAuth: []
    patch:
      tags:
        - User
      summary: Actualizar información de un usuario
      description: Actualiza la información de un usuario específico identificado por su UUID, cumpliendo con el requisito REQ-RYA-U-02.
      operationId: updateUser
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del usuario que se desea actualizar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Usuario actualizado satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "GTU04-GTU05-1", message: "Error en la solicitud al tratar de actualizar un usuario."}
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "GTU04-GTU05-2", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "GTU04-GTU05-3", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "GTU04-GTU05-4", message: "No se encontró el usuario especificado para actualizar."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-02", message: "Error inesperado al tratar de actualizar el usuario."}
      security:
        - bearerAuth: []
    delete:
      tags:
        - User
      summary: Eliminar un usuario existente (soft delete)
      description: Elimina un usuario existente identificado por su UUID de manera suave, cambiándole el estado a "INACTIVE".
      operationId: deleteUser
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuario cambiado de estado a "INACTIVE" satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Usuario eliminado satisfactoriamente."
                  content:
                    default: null
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "deleteUsers401", message: "No está autorizado para eliminar un usuario."}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "deleteUsers403", message: "No tiene permisos para eliminar un usuario."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: { code: "ERRIN-23", message: "Error inesperado al tratar de eliminar el usuario." }
      security:
        - bearerAuth: []
  /api/users/{user_uuid}/favorites:
    get:
      tags:
        - User
      summary: Obtener alojamientos favoritos de un usuario
      description: Obtiene una lista de alojamientos marcados como favoritos por un usuario específico.
      operationId: getUserFavorites
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del usuario cuyos alojamientos favoritos se desean obtener.
      responses:
        '200':
          description: Alojamientos favoritos obtenidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamientos favoritos obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LodgingDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "OPC10-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "OPC10-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Usuario no encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "RYA04-3", message: "No se encontró el usuario especificado para obtener sus alojamientos favoritos."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-03", message: "Error inesperado al tratar de obtener los alojamientos favoritos."}
      security:
        - bearerAuth: []
  /api/users/{user_uuid}/recommendations:
    get:
      tags:
        - User
      summary: Obtener recomendaciones de alojamientos para un usuario
      description: Obtiene una lista de alojamientos recomendados para un usuario específico, basándose en sus preferencias y comportamiento.
      operationId: getUserRecommendations
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del usuario para el cual se desean obtener las recomendaciones de alojamientos.
      responses:
        '200':
          description: Lista de alojamientos recomendados obtenida satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamientos recomendados obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LodgingDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "OPC10-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "OPC10-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Usuario no encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "OPC10-3", message: "No se encontró el usuario especificado para obtener sus recomendaciones."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-04", message: "Error inesperado al tratar de obtener las recomendaciones."}
      security:
        - bearerAuth: []
  /api/users/{user_uuid}/password:
    patch:
      tags:
        - User
      summary: Cambiar la contraseña de un usuario
      description: Permite a un usuario cambiar su contraseña actual, cumpliendo con el requisito REQ-GTU-06.
      operationId: updateUserPassword
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del usuario que desea cambiar su contraseña.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                codeOrPassword:
                  type: string
                  example: Pass123
      responses:
        '200':
          description: Contraseña restablecida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Contraseña restablecida satisfactoriamente."
                  content:
                    default: null
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "GTU06-GTU07-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "GTU06-GTU07-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error de validación."
                  content:
                    default: {code: "GTU06-GTU07-3", message: "El código o la contraseña actual enviada no coincide."}
      security:
        - bearerAuth: []
...
components:
  schemas:
    UserDTO:
      description: Esquema que representa un usuario para transferencia de datos.
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          format: number
        role:
          type: string
          enum: [USER, HOST]
        birthDate:
          type: string
          format: date
      required:
        - name
        - email
        - role
        - birthDate
...
    UserUpdateDTO:
      description: Esquema que representa los datos actualizables de un usuario para transferencia de datos.
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          format: number
        profilePicture:
          type: string
          format: url
          example: "https://example.com/profile.jpg"
        description:
          type: string
        documents:
          type: array
          items:
            type: string
            format: url
            example: "https://example.com/document.pdf"
    LodgingDTO:
      description: Esquema que representa un alojamiento para transferencia de datos.
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        details:
          type: object
          properties:
            location:
              type: object
              properties:
                city:
                  type: string
                address:
                  type: string
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
            price:
              type: number
              format: float
            services:
              type: array
              items:
                type: string
            pictures:
              type: array
              items:
                type: string
                format: url
                example: "https://example.com/image.jpg"
      required:
        - title
        - description
        - details
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT